requirements.txt
celery==5.4.0
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==24.2
django-timezone-field==6.1.0
django==5.0.6
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
eventlet==0.36.1
greenlet==3.0.3
psycopg2==2.9.9
pyjwt==2.8.0
python-crontab==3.0.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
tzdata==2024.1
